Move instruction groups into their own file
