+++ src/include/simeng/arch/riscv/Instruction.hh
@@ -5,6 +5,7 @@
 
 #include "simeng/BranchPredictor.hh"
 #include "simeng/Instruction.hh"
+#include "simeng/arch/riscv/InstructionGroups.hh"
 
 namespace simeng {
 namespace arch {
@@ -21,36 +22,6 @@ const uint8_t FLOAT = 1;
 /** The system registers. */
 const uint8_t SYSTEM = 2;
 }  // namespace RegisterType
-
-/** The IDs of the instruction groups for RISCV instructions. */
-namespace InstructionGroups {
-const uint16_t INT = 0;
-const uint16_t INT_SIMPLE = 1;
-const uint16_t INT_SIMPLE_ARTH = 2;
-const uint16_t INT_SIMPLE_CMP = 3;
-const uint16_t INT_SIMPLE_LOGICAL = 4;
-const uint16_t INT_SIMPLE_SHIFT = 5;
-const uint16_t INT_MUL = 6;
-const uint16_t INT_DIV = 7;
-const uint16_t LOAD_INT = 8;
-const uint16_t STORE_INT = 9;
-const uint16_t LOAD = 10;
-const uint16_t STORE = 11;
-const uint16_t BRANCH = 12;
-}  // namespace InstructionGroups
-
-#define NUM_GROUPS 13
-
-const std::unordered_map<uint16_t, std::vector<uint16_t>> groupInheritance = {
-    {InstructionGroups::INT,
-     {InstructionGroups::INT_SIMPLE, InstructionGroups::INT_MUL,
-      InstructionGroups::INT_DIV}},
-    {InstructionGroups::INT_SIMPLE,
-     {InstructionGroups::INT_SIMPLE_ARTH, InstructionGroups::INT_SIMPLE_CMP,
-      InstructionGroups::INT_SIMPLE_LOGICAL,
-      InstructionGroups::INT_SIMPLE_SHIFT}},
-    {InstructionGroups::LOAD, {InstructionGroups::LOAD_INT}},
-    {InstructionGroups::STORE, {InstructionGroups::STORE_INT}}};
 
 /** A struct holding user-defined execution information for a aarch64
  * instruction. */
